cloud_function/requirements.txt

functions-framework
google-cloud-storage
google-cloud-pubsub
google-genai               
google-cloud-logging
# Database
langchain

langchain-google-vertexai
langchain-text-splitters


cloud-sql-python-connector[asyncpg]
pgvector
langchain-core
asyncpg

# Document Processing
pypdf                      
pymupdf
python-docx
openpyxl

# Text Processing

nltk

# Utilities

tenacity                   
httpx                     
google-cloud-secret-manager

langgraph